Experimenting with support for weapon attachments that are spawned and rendered separately.
Test with M4A1 CQB. Mostly works, but has draw order issues because the hands and first-person weapon are drawn in a different pass than everything else. I can't find any UC code that controls this, so it's likely controlled with native code. It may not be possible to render first-person attachments in the same pass as the weapon they're attached to.